+++ dataforge-workspace/src/jvmTest/kotlin/space/kscience/dataforge/workspace/SimpleWorkspaceTest.kt
@@ -63,7 +63,7 @@ class SimpleWorkspaceTest {
         }
 
         val filterOne by task<Int> {
-            workspace.data.selectOne<Int>("myData[12]")?.let { source ->
+            workspace.data.getByType<Int>("myData[12]")?.let { source ->
                 data(source.name, source.map { it })
             }
         }
@@ -111,23 +111,23 @@ class SimpleWorkspaceTest {
         val sum by task<Int> {
             workspace.logger.info { "Starting sum" }
             val res = from(square).foldToData(0) { l, r ->
-                l + r.await()
+                l + r.value
             }
             data("sum", res)
         }
 
         val averageByGroup by task<Int> {
-            val evenSum = workspace.data.filterIsInstance<Int> { name, _ ->
+            val evenSum = workspace.data.filterByType<Int> { name, _ ->
                 name.toString().toInt() % 2 == 0
             }.foldToData(0) { l, r ->
-                l + r.await()
+                l + r.value
             }
 
             data("even", evenSum)
-            val oddSum = workspace.data.filterIsInstance<Int> { name, _ ->
+            val oddSum = workspace.data.filterByType<Int> { name, _ ->
                 name.toString().toInt() % 2 == 1
             }.foldToData(0) { l, r ->
-                l + r.await()
+                l + r.value
             }
             data("odd", oddSum)
         }
@@ -143,7 +143,7 @@ class SimpleWorkspaceTest {
         }
 
         val customPipe by task<Int> {
-            workspace.data.filterIsInstance<Int>().forEach { data ->
+            workspace.data.filterByType<Int>().forEach { data ->
                 val meta = data.meta.toMutableMeta().apply {
                     "newValue" put 22
                 }
@@ -159,7 +159,7 @@ class SimpleWorkspaceTest {
     fun testWorkspace() {
         runBlocking {
             val node = workspace.runBlocking("sum")
-            val res = node.dataSequence().single()
+            val res = node.traverse().single()
             assertEquals(328350, res.await())
         }
     }
@@ -169,7 +169,7 @@ class SimpleWorkspaceTest {
     fun testMetaPropagation() {
         runBlocking {
             val node = workspace.produce("sum") { "testFlag" put true }
-            val res = node.dataSequence().single().await()
+            val res = node.traverse().single().await()
         }
     }
 
@@ -192,7 +192,7 @@ class SimpleWorkspaceTest {
     fun testFilter() {
         runBlocking {
             val node = workspace.produce("filterOne")
-            assertEquals(12, node.dataSequence().first().await())
+            assertEquals(12, node.traverse().first().await())
         }
     }
 }